+++ common/src/main/java/me/pandadev/fallingtrees/mixin/BlockStateBaseMixin.java
@@ -20,14 +20,9 @@ public abstract class BlockStateBaseMixin {
 
 	@Inject(method = "getDestroyProgress", at = @At("RETURN"), cancellable = true)
 	public void getDestroyProgress(Player player, BlockGetter level, BlockPos pos, CallbackInfoReturnable<Float> cir) {
-		if (TreeUtils.isLog(this.getBlock())) {
-			TreeCache cache = TreeCache.getOrCreateCache("tree_breaking", pos, level, player);
-			if (FallingTrees.getServerConfig().tree_mining_speed_by_log_amount && TreeUtils.shouldTreeFall(player) &&
-					!cache.isTreeSizeToBig() && cache.isTree())
-				cir.setReturnValue(cir.getReturnValue() /
-						((Math.min(FallingTrees.getServerConfig().tree_mining_speed_max_log_limit,
-								cache.getLogAmount())-1) *
-								FallingTrees.getServerConfig().tree_mining_speed_multiplier + 1));
-		}
+		TreeCache cache = TreeCache.getOrCreateCache("tree_breaking", pos, level, player);
+		if (cache != null && FallingTrees.getServerConfig().tree_mining_speed_by_log_amount && TreeUtils.shouldTreeFall(player) && !cache.isTreeSizeToBig())
+			cir.setReturnValue(cir.getReturnValue() / ((Math.min(FallingTrees.getServerConfig().tree_mining_speed_max_log_limit, cache.getLogAmount())-1) *
+					FallingTrees.getServerConfig().tree_mining_speed_multiplier + 1));
 	}
 }