+++ common/src/main/java/me/pandadev/fallingtrees/mixin/MultiPlayerGameModeMixin.java
@@ -1,24 +1,20 @@
 package me.pandadev.fallingtrees.mixin;
 
-import com.sun.source.tree.Tree;
 import me.pandadev.fallingtrees.network.BreakTreePacket;
 import me.pandadev.fallingtrees.tree.TreeCache;
-import me.pandadev.fallingtrees.tree.TreeUtils;
+import me.pandadev.fallingtrees.trees.DefaultTreeType;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.multiplayer.MultiPlayerGameMode;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.level.block.state.BlockState;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.Redirect;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(MultiPlayerGameMode.class)
@@ -33,18 +29,14 @@ public class MultiPlayerGameModeMixin {
 		Player player = this.minecraft.player;
 		if (level != null && player != null) {
 			TreeCache cache = TreeCache.getOrCreateCache("tree_breaking", pos, level, player);
-			if (cache.isTreeSizeToBig())
+			if (cache == null || cache.isTreeSizeToBig())
 				return;
-			BlockState state = level.getBlockState(pos);
-			if (TreeUtils.isLog(state.getBlock())) {
-				BreakTreePacket.sendToServer(pos, player);
-			}
+			BreakTreePacket.sendToServer(pos, player);
 		}
 	}
 
 	@Unique
 	boolean isShifting = true;
-
 	@Inject(method = "startDestroyBlock", at = @At("HEAD"))
 	public void startDestroyBlock(BlockPos loc, Direction face, CallbackInfoReturnable<Boolean> cir) {
 		Player player = this.minecraft.player;
@@ -59,7 +51,7 @@ public class MultiPlayerGameModeMixin {
 	public void continueDestroyBlock(BlockPos posBlock, Direction directionFacing, CallbackInfoReturnable<Boolean> cir) {
 		Player player = this.minecraft.player;
 		Level level = this.minecraft.level;
-		if (level != null && TreeUtils.isLog(level.getBlockState(posBlock).getBlock()) && player != null) {
+		if (level != null && DefaultTreeType.isLog(level.getBlockState(posBlock).getBlock()) && player != null) {
 			if (player.isCrouching() != isShifting) {
 				isShifting = player.isCrouching();
 				this.destroyProgress = 0;