+++ common/src/main/java/me/pandadev/fallingtrees/tree/TreeCache.java
@@ -1,20 +1,18 @@
 package me.pandadev.fallingtrees.tree;
 
-import com.sun.source.tree.Tree;
 import me.pandadev.fallingtrees.FallingTrees;
 import me.pandadev.fallingtrees.config.ServerConfig;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.BlockGetter;
-import net.minecraft.world.level.Level;
-import net.minecraft.world.level.block.state.BlockState;
 import org.joml.Vector3i;
 
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 
-public record TreeCache(Vector3i pos, List<BlockPos> treeBlocks, BlockGetter level) {
+public record TreeCache(Vector3i pos, List<BlockPos> blocks, BlockGetter level, TreeType treeType) {
 	public static Map<String, Map<Integer, TreeCache>> CACHES = new HashMap<>();
 
 	public static Map<Integer, TreeCache> getCaches(String name) {
@@ -26,22 +24,18 @@ public record TreeCache(Vector3i pos, List<BlockPos> treeBlocks, BlockGetter lev
 
 	public int getLogAmount() {
 		int amount = 0;
-		for (BlockPos treeBlock : treeBlocks) {
-			if (TreeUtils.isLog(level.getBlockState(treeBlock).getBlock())) amount++;
+		for (BlockPos treeBlockPos : blocks) {
+			if (treeType.blockChecker(treeBlockPos, level)) amount++;
 		}
 		return amount;
 	}
 
 	public boolean isTreeSizeToBig() {
 		if (FallingTrees.getServerConfig().tree_limit.tree_limit_method.equals(ServerConfig.TreeLimit.LimitMethodEnum.BLOCKS)) {
-			return treeBlocks.size() > FallingTrees.getServerConfig().tree_limit.tree_size_limit;
+			return blocks.size() > FallingTrees.getServerConfig().tree_limit.tree_size_limit;
 		}
-		return treeBlocks.stream().filter(blockPos ->
-				TreeUtils.isLog(level.getBlockState(blockPos).getBlock())).count() > FallingTrees.getServerConfig().tree_limit.tree_size_limit;
-	}
-
-	public boolean isTree() {
-		return true;
+		return blocks.stream().filter(blockPos ->
+				treeType.blockChecker(blockPos, level)).count() > FallingTrees.getServerConfig().tree_limit.tree_size_limit;
 	}
 
 	public static TreeCache getOrCreateCache(String cacheName, BlockPos pos, BlockGetter level, Player player) {
@@ -52,7 +46,14 @@ public record TreeCache(Vector3i pos, List<BlockPos> treeBlocks, BlockGetter lev
 	}
 
 	public static TreeCache createCache(String cacheName, BlockPos pos, BlockGetter level, Player player) {
-		TreeCache cache = new TreeCache(new Vector3i(pos.getX(), pos.getY(), pos.getZ()), TreeUtils.getTreeBlocks(pos, level), level);
+		Optional<TreeType> treeType = TreeTypeRegistry.getTreeType(pos, level);
+		if (treeType.isEmpty())
+			return null;
+		List<BlockPos> blocks = treeType.get().blockDetectionAlgorithm(pos, level);
+		if (blocks.size() <= 1)
+			return null;
+
+		TreeCache cache = new TreeCache(new Vector3i(pos.getX(), pos.getY(), pos.getZ()), blocks, level, treeType.get());
 		getCaches(cacheName).put(player.getId(), cache);
 		return cache;
 	}