+++ common/src/main/java/me/pandadev/fallingtrees/tree/TreeType.java
@@ -1,11 +1,17 @@
 package me.pandadev.fallingtrees.tree;
 
-import java.util.concurrent.locks.Condition;
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.level.block.state.BlockState;
 
-public class TreeType {
-	public final TreeCondition conditions;
+import java.util.List;
+import java.util.Map;
 
-	public TreeType(TreeCondition conditions) {
-		this.conditions = conditions;
+public abstract class TreeType {
+	public abstract boolean blockChecker(BlockPos pos, BlockGetter level);
+	public abstract List<BlockPos> blockDetectionAlgorithm(BlockPos pos, BlockGetter level);
+
+	public boolean extraBlockRequirement(Map<BlockPos, BlockState> blocks, BlockGetter level) {
+		return true;
 	}
 }