+++ Changelog.md
@@ -2,6 +2,8 @@
 * **Added**
   * One block mining indicator
   * Food exhaustion limit
+  * Limit for mining speed
+  * Added support for Bamboo and Cactus
 * **Changes**
   * Updated Networking to try and fix crashes when breaking logs.
   * Improved block rendering so blocks keep tint and hopefully fix more crashes.